requirements.txt
pandas==1.4.3
seaborn==0.11.2
matplotlib==3.5.2
streamlit==1.39.0
Pillow~=9.2.0
numpy~=1.23.1